[CHORE] record 앱 분리
