+++ game/models/game.py
@@ -6,11 +6,3 @@ class Game(models.Model):
 
     class Meta:
         db_table = 'games'
-
-class Record(models.Model):
-    game = models.ForeignKey('Game', on_delete=models.CASCADE)
-    trash = models.ForeignKey('Trash', on_delete=models.CASCADE)
-    is_answer = models.BooleanField(null=True)
-
-    class Meta:
-        db_table = 'records'
